chore: delete unnecessary code refactor: move actions and state to proper store
