+++ src/store/lessons/contentStore.js
@@ -1,23 +1,28 @@
 import { defineStore } from "pinia";
 
+import { getContentsRelationByDays } from "@/requests/daysRequest.js";
 import { getMD } from "@/requests/contentRequest.js";
 
 export const useContentStore = defineStore("content", {
     actions: {
         async getText(path) {
-            this.lesson = await getMD(path);
+            const content = await getMD(path);
+            this.lesson = content;
 
             return this.lesson;
         },
-        setUuid(uuid) {
-            this.uuid = uuid;
+        async getByContent(uuid = "") {
+            const data = await getContentsRelationByDays(uuid);
+
+            this.result = [];
+            this.result = data.node;
         }
     },
     state: () => {
         return {
             lesson    : "",
             contIndex : 0,
-            uuid      : ""
+            result    : []
         };
     }
 });
