+++ src/store/lessons/daysStore.js
@@ -1,12 +1,13 @@
 import { defineStore } from "pinia";
 
-import { getContentsRelationByDays } from "@/requests/daysRequest.js";
 import { getSprintByUuid } from "@/requests/sprintsRequest.js";
+import { useSprintsStore } from "./sprintsStore.js";
 
 export const useDaysStore = defineStore("days", {
     actions: {
         async getDaysByModule() {
-            const data = await getSprintByUuid(this.sprintUuid);
+            const sprintsStore = useSprintsStore();
+            const data = await getSprintByUuid(sprintsStore.sprintUuid);
 
             this.days = data.node.map(value => {
                 return {
@@ -14,22 +15,11 @@ export const useDaysStore = defineStore("days", {
                     desc : value.node.desc
                 };
             });
-        },
-        async getByContent(uuid = "") {
-            const data = await getContentsRelationByDays(uuid);
-
-            this.result = [];
-            this.result = data.node;
-        },
-        setDaysBySprintUuid(uuid) {
-            this.sprintUuid = uuid;
         }
     },
     state: () => {
         return {
-            days       : [],
-            result     : [],
-            sprintUuid : ""
+            days: []
         };
     }
 });
